refactor: splitting post into layout and page
