+++ src/routes/api/posts/[slug]/+server.ts
@@ -1,6 +1,6 @@
 import { PUBLIC_CMS_URL } from '$env/static/public';
 import { json } from '@sveltejs/kit';
-import { compileMarkdown, estimateReadingTime } from '$lib/utils';
+import { estimateReadingTime } from '$lib/utils';
 import qs from 'qs';
 import type { ICMSData, IPost } from '$lib/ts';
 
@@ -24,7 +24,6 @@ export const GET = async ({ fetch, params: { slug } }) => {
 			.then(res => res.data.at(0) as IPost)
 			.then(async res => ({
 				...res,
-				content: await compileMarkdown(res.content),
 				readingTime: estimateReadingTime(res.content),
 				tags: res.tags.sort((a, b) => a.value.localeCompare(b.value))
 			}))
