+++ src/routes/blog/[slug]/+page.server.ts
@@ -1,21 +1,9 @@
-import qs from 'qs';
-import type { ICMSData, IPost } from '$lib/ts';
+import { compileMarkdown } from '$lib/utils';
 
-export const load = async ({ fetch, params: { slug } }) => {
-	const post = await fetch(`/api/posts/${slug}`).then<IPost>(res => res.json());
+export const load = async ({ parent }) => {
+	const data = await parent();
 
-	const query = {
-		related: {
-			tags: post.tags.map(t => t.value)
-		}
-	};
+	const compiledContent = await compileMarkdown(data.post.content);
 
-	const relatedPosts = await fetch(
-		`/api/posts/${slug}/related?${qs.stringify(query.related, {
-			encodeValuesOnly: true,
-			arrayFormat: 'repeat'
-		})}`
-	).then<ICMSData<IPost[]>>(res => res.json());
-
-	return { post, relatedPosts };
+	return { compiledContent };
 };
