+++ svelte.config.js
@@ -3,13 +3,12 @@ import adapter from '@sveltejs/adapter-vercel';
 import preprocess from 'svelte-preprocess';
 import seqPreprocessor from 'svelte-sequential-preprocessor';
 import { preprocessThrelte } from '@threlte/preprocess';
-import { mdsvex } from 'mdsvex';
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
 	// Consult https://kit.svelte.dev/docs/integrations#preprocessors
 	// for more information about preprocessors
-	preprocess: seqPreprocessor([vitePreprocess(), preprocess(), preprocessThrelte(), mdsvex()]),
+	preprocess: seqPreprocessor([vitePreprocess(), preprocess(), preprocessThrelte()]),
 
 	kit: {
 		adapter: adapter()
